+++ api_gen/api_gen.py
@@ -1,49 +1,21 @@
 #!/usr/bin/env python3
 
-import re
-import os
-import sys
-from glob import iglob
-import urllib.parse
 import html.parser
 import itertools
+import os
+import re
+import shutil
+import subprocess
+import urllib.parse
 
-import inflection
 from bs4 import BeautifulSoup
+import inflection
+import requests
 
-# Unless you're really going to hack on the API generator, don't hammer
-# the developer site needlessly. Currently, mirroring the whole site
-# results in about 26MB on disk. I may eventually do some more intelligent
-# scraping, downloading, and disk caching to avoid this wget.
 
-# That said, here is my usual process:
-# $ wget -mk https://developer.zendesk.com/rest_api/docs/core/introduction
-# $ cp -R developer.zendesk.com developer.zendesk.com.orig
-# $ ./clean.py
-# $ cp -R developer.zendesk.com developer.zendesk.com.cleaned
-#
-# For each patch.* file
-# cd developer.zendesk.com
-# $ patch -p1 < ../patch.file
-#
-# $ cd ..
-# $ ./api_gen.py
-
-# If I find that more patching is needed to get what I want, then:
-# $ vim edit whatever
-# $ diff -r -u developer.zendesk.com developer.zendesk.com.cleaned > api_gen.patch.new
-# If that all checks out, then api_gen.patch.new becomes the new api_gen.patch
-
-# Then, once all is well:
-# cp zdesk_api.py ../zdesk/
+# See README.md for documentation.
 
 html_parser = html.parser.HTMLParser()
-
-skip_files=[
-    'introduction',
-    'changes_roadmap',
-    'side_loading',
-    ]
 
 api_actions = [
     'autocomplete',
@@ -56,6 +28,7 @@ api_actions = [
     'exports',
     'import',
     'imports',
+    'logout',
     'make',
     'mark',
     'me',
@@ -73,30 +46,143 @@ api_actions = [
 with open('api_template.py', 'r') as template_file:
     template = template_file.read()
 
-#os.chdir('old.developer.zendesk.com/documentation/rest_api')
-#os.chdir('developer.zendesk.com/rest_api/docs/core')
+# os.chdir('old.developer.zendesk.com/documentation/rest_api')
+# os.chdir('developer.zendesk.com/rest_api/docs/core')
+
+zen_url = 'https://developer.zendesk.com'
+
+docpages = {
+    'core': '/rest_api/docs/core/introduction',
+    'webportal': '/rest_api/docs/web-portal/webportal_introduction',
+    'hc': '/rest_api/docs/help_center/introduction',
+    'zopim': '/rest_api/docs/zopim/introduction',
+    'voice': '/rest_api/docs/voice-api/voice',
+    'nps': '/rest_api/docs/nps-api/introduction',
+}
+
+skippages = [
+    '/rest_api/docs/core/introduction',
+    '/rest_api/docs/core/getting_started',
+    '/rest_api/docs/core/api_changes',
+    '/rest_api/docs/core/restrictions',
+    '/rest_api/docs/help_center/introduction',
+    '/rest_api/docs/zopim/introduction',
+    '/rest_api/docs/zopim/restrictions',
+    '/rest_api/docs/zopim/changes_roadmap',
+    '/rest_api/docs/web-portal/webportal_introduction',
+    '/rest_api/docs/nps-api/introduction',
+]
+
+skipfiles = [
+    'core_introduction',
+    'webportal_webportal_introduction',
+    'hc_introduction',
+    'zopim_introduction',
+    'nps_introduction',
+]
+
+apidocs = 'apidocs'
+
+if os.path.isdir(apidocs):
+    shutil.rmtree(apidocs)
+
+if not os.path.isdir(apidocs + '_orig'):
+    os.makedirs(apidocs + '_orig')
+
+os.chdir(apidocs + '_orig')
+
+for category in docpages:
+    filename = category + '_' + os.path.basename(docpages[category])
+
+    if os.path.isfile(filename):
+        with open(filename, 'rb') as fin:
+            content = fin.read()
+    else:
+        req = requests.get(zen_url + docpages[category])
+        content = req.content
+
+        with open(filename, 'wb') as fout:
+            fout.write(content)
+
+    soup = BeautifulSoup(content, "html.parser")
+    sidenav = soup(attrs={'class': 'docs-sidenav'})[0]
+
+    for a in sidenav.find_all('a'):
+        link = a.attrs['href']
+        filename = category + '_' + os.path.basename(link)
+
+        if link[0] == '#':
+            continue
+        if link in skippages:
+            continue
+        if os.path.isfile(filename):
+            continue
+
+        req = requests.get(zen_url + link)
+        text = BeautifulSoup(req.content, "html.parser")
+
+        pretty_text = text.prettify()
+        if not pretty_text:
+            # In the case of prettify killing the whole page
+            pretty_text = text
+        pretty = (pretty_text)
+
+        with open(category + '_' + os.path.basename(link), 'w') as fout:
+            fout.write(pretty)
+
+doc_files = os.listdir()
+os.chdir('..')
+patchfiles = os.listdir('patches')
+shutil.copytree(apidocs + '_orig', apidocs)
+os.chdir(apidocs)
+
+for patchfile in patchfiles:
+    try:
+        sp = subprocess.check_output([
+            'patch', '-p1', patchfile,
+            os.path.join('..', 'patches', patchfile)])
+    except subprocess.CalledProcessError as e:
+        print('Failed to patch {}. Exit {}\n'.format(patchfile, e.returncode))
+        print('Output was:\n')
+        print(e.output.decode())
+
+def code_pre_docanchor(tag):
+    class_ = tag.attrs.get('class', ())
+    return (tag.name == 'code' or tag.name == 'pre' or
+            (tag.name == 'a' and 'doc-anchor-link' in class_))
 
 api_items = {}
 duplicate_api_items = {}
-for doc_file in iglob(os.path.join('developer.zendesk.com', 'rest_api', 'docs', '*', '*')):
-    if '.html' in doc_file or '.orig' in doc_file or os.path.split(doc_file)[1] in skip_files:
+for doc_file in doc_files:
+    if '.orig' in doc_file or doc_file in skipfiles:
         continue
 
-    with open(doc_file, 'r') as doc:
-        soup = BeautifulSoup(doc)
+    with open(doc_file) as doc:
+        soup = BeautifulSoup(doc, "html.parser")
 
-    for code in soup.find_all(['code', 'pre']):
-        text = code.get_text()
-        #if re.search(r'<code>(OPTIONS|GET|HEAD|POST|PUT|DELETE|TRACE|CONNECT) .*\.json', line):
+    for tag in soup.find_all(code_pre_docanchor):
+        if tag.name == 'a':
+            last_doc_anchor = tag['href']
+            continue
+
+        text = tag.get_text()
+        # if re.search(
+        #    r'<code>(OPTIONS|GET|HEAD|POST|PUT|DELETE|TRACE|CONNECT) .*\.json'
+        # , line):
         #    line = re.sub(r'(<p>|<code>|</p>|</code>|/api/v2)', '', line)
         #    line = re.sub('&#123;', '{', line)
         #    line = re.sub('&#125;', '}', line)
         #    print(line, end='')
-        match = re.match(r'\s*(OPTIONS|GET|HEAD|POST|PUT|DELETE|TRACE|CONNECT) (.*)', text)
+        match = re.match(
+            r'\s*(OPTIONS|GET|HEAD|POST|PUT|DELETE|TRACE|CONNECT) (.*)', text)
         if match:
+            cat = os.path.basename(doc_file.split('_')[0])
+            page = doc_file.replace(cat + '_', '')
+            docpage = zen_url + os.path.dirname(docpages[cat]) + '/' + page
+
             is_singular = False
             api_item = {}
-            api_item['docpage'] = doc_file
+            api_item['docpage'] = docpage + last_doc_anchor
             api_item['path_params'] = []
             api_item['opt_path_params'] = []
             api_item['opt_path'] = ''
@@ -106,12 +192,13 @@ for doc_file in iglob(os.path.join('developer.zendesk.com', 'rest_api', 'docs', 
             api_item['method'] = match.group(1)
             path = match.group(2)
 
-            # parse into a url object for easily accessing the parts of the whole url
+            # parse into a url object for easily accessing the parts of the url
             url = urllib.parse.urlsplit(path)
 
             api_item['path'] = url.path
 
-            # Split the path and remove the first (always empty) item and /api/v2
+            # Split the path and remove the first (always empty) item and
+            # /api/v2
             param_indexes = []
             path_parts = url.path.split('/')
 
@@ -128,10 +215,7 @@ for doc_file in iglob(os.path.join('developer.zendesk.com', 'rest_api', 'docs', 
             path_parts.reverse()
             for i, path_part in enumerate(path_parts):
                 part, ext = os.path.splitext(path_part)
-                if ext != '':
-                    if ext != '.json':
-                        print('Non-JSON endpoint encountered!!')
-                    path_parts[i] = part
+                path_parts[i] = part
 
                 if part.startswith('{'):
                     api_item['path_params'].append(part[1:-1])
@@ -150,8 +234,8 @@ for doc_file in iglob(os.path.join('developer.zendesk.com', 'rest_api', 'docs', 
 
             expanded_parts = []
             [expanded_parts.extend(part.split('_')) for part in path_parts]
-            has_action = True in [action in expanded_parts
-                    for action in api_actions]
+            has_action = True in [
+                action in expanded_parts for action in api_actions]
 
             query_params = urllib.parse.parse_qsl(url.query)
             for query_items in query_params:
@@ -174,12 +258,8 @@ for doc_file in iglob(os.path.join('developer.zendesk.com', 'rest_api', 'docs', 
                 elif (
                     api_item['method'] == 'GET' and
                     len(api_item['path_params']) == 0
-                   ):
+                ):
                     name = name + '_list'
-            else:
-                # one hard corner case with 'me' and 'sessions delete'
-                if 'me' in expanded_parts and api_item['method'] == 'DELETE':
-                    name = name + '_delete'
 
             api_item['path_params'].reverse()
             api_item['query_params'].reverse()
@@ -209,52 +289,141 @@ for name in names:
                 dupe['path_params'], item['path_params'])] and
             False not in [i == j for i, j in itertools.zip_longest(
                 sorted(dupe['query_params']), sorted(item['query_params']))]
-           ):
+        ):
             # Everything is the same, so discard this duplicate
-            content += "    # Duplicate API endpoint discarded: {} from {}\n".format(name, dupe['docpage'])
+            content += \
+                "    # Duplicate API endpoint discarded:\n"\
+                "    # {} from\n"\
+                "    # {}\n\n".format(
+                    name, dupe['docpage'])
         elif (
             dupe['method'] == item['method'] and
             False not in [i == j for i, j in itertools.zip_longest(
-                sorted(dupe['query_params']), sorted(item['query_params']))]
-            ):
-            # Only the path parameters differ, so we have optional arguments
+                sorted(dupe['query_params']), sorted(item['query_params']))]):
+            # Only the path parameters differ, so we have optional or ambiguous
+            # arguments
 
             # Common parameters are all required
             required = set(dupe['path_params']) & set(item['path_params'])
 
             # Optional parameters are only in one endpoint
-            optional = list((set(dupe['path_params']) | set(item['path_params'])) - required)
+            optional = list((
+                set(dupe['path_params']) | set(item['path_params'])) -
+                required)
 
-            if len(set(item['path_params']) - required) == 0:
-                # The item is the base function and the dupe has the optional arguments
-                # Just need to add the optional arguments
+            if (len(set(item['path_params']) - required) == 0 and
+                    len(optional) > 0):
+                # The item is the base function and the dupe has the optional
+                # arguments. Just need to add the optional arguments
                 item['opt_path_params'] = optional
                 item['opt_path'] = dupe['path']
-            elif len(set(dupe['path_params']) - required) == 0:
-                # The dupe is the base function and the item has the optional arguments
-                # Need to swap the dupe with the item and then add the optional arguments
+            elif (len(set(dupe['path_params']) - required) == 0 and
+                    len(optional) > 0):
+                # The dupe is the base function and the item has the
+                # optional arguments
+                # Need to swap the dupe with the item and then add the
+                # optional arguments
                 dupe_path = dupe['path']
                 item = dupe
                 item['opt_path_params'] = optional
                 item['opt_path'] = dupe_path
             else:
-                content += "    # Duplicate ambiguous API endpoint: {} from {}\n".format(name, dupe['docpage'])
-                continue
+                # this is ambiguous. If the parameters are in the same place,
+                # then they're actually interchangeable, and we just need a
+                # better name for them.
+                handled = True
+                new_path = ''
+                new_path_params = []
+
+                # compare each of the path parts to see specifically what is
+                # different.
+                for i, j in itertools.zip_longest(
+                        item['path'].split('/'), dupe['path'].split('/')):
+                    if i == j:
+                        # everything is the same up to here. keep building a
+                        # new, common path.
+                        if i:
+                            new_path += '/' + i
+                    else:
+                        # the paths are different. look at how they are
+                        # different.
+                        ipart, iext = os.path.splitext(i)
+                        jpart, jext = os.path.splitext(j)
+                        if ipart == jpart and iext != jext:
+                            # These are legit dupes that only differ by
+                            # the extension at the end.
+                            content += \
+                                "    # Duplicate API endpoint differs "\
+                                "only by extension:\n"\
+                                "    # {} from\n"\
+                                "    # {}\n\n".format(
+                                    name, dupe['docpage'])
+                            handled = True
+                            break
+
+                        if (ipart.startswith('{') and jpart.startswith('{') and
+                                iext != jext):
+                            # The parameters are different, but in the same
+                            # place and have a different extension. So this
+                            # actually needs a new method.
+                            # e.g. /thing/{id} vs /thing/{name}.json
+                            new_name = name + '_by_' + jpart.strip('{}')
+                            if new_name not in api_items:
+                                api_items[new_name] = dupe
+                                handled = True
+                                break
+
+                        if (ipart.startswith('{') and jpart.startswith('{') and
+                                iext == jext):
+                            # The parameters are in the same place and have the
+                            # same extension. Combine these parameters as they
+                            # are basically interchangeable.
+                            # e.g. /thing/{id} vs /thing/{name} becomes
+                            #      /thing/{id_or_name}
+                            new_param = i.strip('{}' + iext) \
+                                + '_or_' + j.strip('{}' + jext)
+                            new_path_params.append(new_param)
+                            new_path += '/{' + new_param + '}' + iext
+                            item['path'] = new_path
+                            item['path_params'] = new_path_params
+                            handled = True
+                            break
+
+                if not handled:
+                    content += \
+                        "    # Duplicate ambiguous API endpoint:\n"\
+                        "    # {} from\n"\
+                        "    # {}\n\n".format(name, dupe['docpage'])
         elif (
             dupe['path'] == item['path'] and
             dupe['method'] == item['method'] and
             False not in [i == j for i, j in itertools.zip_longest(
-                dupe['path_params'], item['path_params'])]
-            ):
-                # Only the query parameters differ, so we have optional named arguments
-                item['opt_query_params'] = list(set(item['query_params'].copy() +
-                                               item['opt_query_params'] +
-                                               dupe['query_params']))
+                dupe['path_params'], item['path_params'])]):
+                # Only the query parameters differ, so we have
+                # optional named arguments
+                item['opt_query_params'] = list(
+                    set(item['query_params'].copy() +
+                        item['opt_query_params'] +
+                        dupe['query_params']))
                 item['query_params'] = []
+        elif (
+            dupe['path'] == item['path'] and
+            False not in [i == j for i, j in itertools.zip_longest(
+                dupe['path_params'], item['path_params'])] and
+            False not in [i == j for i, j in itertools.zip_longest(
+                sorted(dupe['query_params']), sorted(item['query_params']))]):
+                # Only the method differs, so indicate that method is ambiguous
+                item['method'] = None
         else:
             should_keep += "\n"
-            should_keep += "    # Duplicate API endpoint that differs: {} from {}\n".format(name, dupe['docpage'])
-            should_keep += "    # Original definition located here:    {} from {}\n".format(name, item['docpage'])
+            should_keep += \
+                "    # Duplicate API endpoint that differs:\n"\
+                "    #  {} from\n"\
+                "    #  {}\n".format(name, dupe['docpage'])
+            should_keep += \
+                "    # Original definition located here:\n"\
+                "    #  {} from\n"\
+                "    #  {}\n".format(name, item['docpage'])
             should_keep += '    # {}\n'.format(name)
             should_keep += '    #    Path: {}\n'.format(dupe['path'])
             should_keep += '    #    Method: {}\n'.format(dupe['method'])
@@ -276,11 +445,7 @@ for name in names:
             should_keep += "\n"
 
 if should_keep:
-    content += '\n'
     content += should_keep
-
-if content:
-    content += '\n'
 
 names = list(api_items.keys())
 names.sort()
@@ -305,10 +470,16 @@ for name in names:
         paramspec += ', '
     argspec = paramspec + queryspec
 
-    if item['method'] in ['POST', 'PUT']:
+    if item['method']:
+        if item['method'] in ['POST', 'PUT']:
+            if argspec:
+                argspec += ', '
+            argspec += 'data'
+    else:
         if argspec:
-            argspec += ', '
-        argspec += 'data'
+            argspec = 'method, ' + argspec + ', data=None'
+        else:
+            argspec = 'method, data=None'
 
     if item['opt_path_params']:
         if argspec:
@@ -318,27 +489,31 @@ for name in names:
     if item['opt_query_params']:
         if argspec:
             argspec += ', '
-        argspec += '=None, '.join([sanitize(q) for q in item['opt_query_params']]) + '=None'
+        argspec += '=None, '.join(
+            [sanitize(q) for q in item['opt_query_params']]) + '=None'
 
     if argspec:
         argspec += ', '
     argspec += '**kwargs'
 
     content += '    def {}(self, {}):\n'.format(name, argspec)
-    content += '        "http://{}"\n'.format( item['docpage'])
+    content += '        "{}"\n'.format(item['docpage'])
     content += '        api_path = "{}"\n'.format(item['path'])
 
     if item['path_params']:
-        content += '        api_path = api_path.format({})\n'.format(path_fmt_args)
+        content += '        api_path = api_path.format({})\n'.format(
+            path_fmt_args)
 
     if item['opt_path']:
         opt_test = ' and '.join(item['opt_path_params'])
         if path_fmt_args:
             path_fmt_args += ', '
-        path_fmt_args = path_fmt_args + ', '.join([p + '=' + p for p in item['opt_path_params']])
+        path_fmt_args = path_fmt_args + \
+            ', '.join([p + '=' + p for p in item['opt_path_params']])
         content += '        if {}:\n'.format(opt_test)
         content += '            api_opt_path = "{}"\n'.format(item['opt_path'])
-        content += '            api_path = api_opt_path.format({})\n'.format(path_fmt_args)
+        content += '            api_path = api_opt_path.format({})\n'.format(
+            path_fmt_args)
 
     if item['query_params'] or item['opt_query_params']:
         content += '        api_query = {}\n'
@@ -356,8 +531,9 @@ for name in names:
         content += '            })\n'
 
     # todo: query may be required
-    #if item['opt_query_params']:
-    #    opt_test = ' or '.join([sanitize(q) for q in item['opt_query_params']])
+    # if item['opt_query_params']:
+    #    opt_test = ' or '.join(
+    #       [sanitize(q) for q in item['opt_query_params']])
     #    content += '        if not ({}):\n'.format(opt_test)
     #    content += '            pass\n'
 
@@ -366,14 +542,18 @@ for name in names:
     if item['query_params'] or item['opt_query_params']:
         content += ', query=api_query'
 
-    if item['method'] != 'GET':
-        content += ', method="{}"'.format(item['method'])
+    if item['method']:
+        if item['method'] != 'GET':
+            content += ', method="{}"'.format(item['method'])
 
-    if item['method'] in ['POST', 'PUT']:
-        content += ', data=data'
+        if item['method'] in ['POST', 'PUT']:
+            content += ', data=data'
+    else:
+        content += ', method=method, data=data'
 
     content += ', **kwargs)\n\n'
 
+os.chdir('..')
+
 with open('zdesk_api.py', 'w') as f:
     f.write(template.format(content))
-